chore: Reorganize the tests and reduce redundant tests
Many tests are covered by the hierarchy lookup test
