+++ ors-test-scenarios/src/test/java/integrationtests/GeoToolsTest.java
@@ -1,8 +1,14 @@
 package integrationtests;
 
+import org.junit.jupiter.api.MethodOrderer;
+import org.junit.jupiter.api.TestInstance;
+import org.junit.jupiter.api.TestMethodOrder;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.testcontainers.containers.GenericContainer;
+import org.testcontainers.junit.jupiter.Testcontainers;
+import org.testcontainers.junit.jupiter.TestcontainersExtension;
 import utils.ContainerInitializer;
 import utils.OrsContainerFileSystemCheck;
 
@@ -11,12 +17,16 @@ import java.io.IOException;
 import static utils.ContainerInitializer.initContainer;
 import static utils.OrsApiHelper.checkAvoidAreaRequest;
 
+@ExtendWith(TestcontainersExtension.class)
+@TestInstance(TestInstance.Lifecycle.PER_METHOD)
+@TestMethodOrder(MethodOrderer.OrderAnnotation.class)
+@Testcontainers(disabledWithoutDocker = true)
 public class GeoToolsTest {
 
     @MethodSource("utils.ContainerInitializer#ContainerTestImageDefaultsImageStream")
     @ParameterizedTest(name = "{0}")
     void testAvoidAreaRequestAndGeoToolsPopulation(ContainerInitializer.ContainerTestImageDefaults targetImage) throws IOException, InterruptedException {
-        GenericContainer<?> container = initContainer(targetImage, true, true);
+        GenericContainer<?> container = initContainer(targetImage, true);
 
         String geoToolsPath;
         if (targetImage.equals(ContainerInitializer.ContainerTestImageDefaults.WAR_CONTAINER))
@@ -26,5 +36,6 @@ public class GeoToolsTest {
         OrsContainerFileSystemCheck.assertDirectoryExists(container, geoToolsPath, false);
         checkAvoidAreaRequest("http://" + container.getHost() + ":" + container.getFirstMappedPort() + "/ors/v2/directions/driving-car/geojson", 200);
         OrsContainerFileSystemCheck.assertDirectoryExists(container, geoToolsPath, true);
+        container.stop();
     }
 }
